requirements.txt
amqp==2.6.1
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.27.0
backcall==0.2.0
billiard==3.6.4.0
celery==4.4.7
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
colorama==0.4.6
decorator==5.1.1
docformatter==1.6.5
executing==2.0.1
fastapi==0.95.0
flower==0.9.7
greenlet==3.0.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.0
humanize==4.8.0
idna==3.4
ipython==8.13.2
jedi==0.19.1
Jinja2==3.1.2
kombu==4.6.11
lxml==4.9.2
lxml-stubs==0.4.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.5
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.7
Pygments==2.17.1
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
python-redis-lock==4.0.0
pytz==2023.3.post1
redis==4.5.4
requests==2.28.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
SQLAlchemy-Utils==0.40.0
sqlparse==0.4.3
stack-data==0.6.3
starlette==0.26.1
toposort==1.10
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.8.0
untokenize==0.1.1
urllib3==1.26.15
uvicorn==0.21.1
vine==1.3.0
wcwidth==0.2.10
xsdata==23.5
flask~=3.1.0